backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
openai==1.3.0
python-dotenv==1.0.0
werkzeug==2.2.3
httpx==0.24.1
beautifulsoup4==4.12.2
sentence-transformers==2.4.0
pandas==2.1.1
matplotlib==3.8.0
redis==5.0.1
langchain==0.1.6
langchain-openai==0.0.5
lxml==4.9.3
fake-useragent==1.4.0
waitress==2.1.2